Extracted vuetify theme into separate settings file
